+++ extensions-contrib/time-min-max/src/main/java/io/druid/query/aggregation/TimestampBufferAggregator.java
@@ -82,7 +82,7 @@ public class TimestampBufferAggregator implements BufferAggregator
   @Override
   public double getDouble(ByteBuffer buf, int position)
   {
-    return buf.getDouble(position);
+    return (double) buf.getLong(position);
   }
 
   @Override
